refactor(#728): extract lookup/validation.py + lookup/artwork.py from the orchestrator
Pure code motion, no behavior change (PR 5 of the LML#722 decomposition). lookup/validation.py takes the Step-3b track validation (filter_results_by_track_validation) and the A4 cached-track safety net (find_library_albums_with_cached_track); lookup/artwork.py takes the Step-4 artwork fetch (fetch_artwork_for_items), the LML#604 trust-and-bind (_bind_resolved_release), the cover/artist/label fallback cascade (_resolve_fallback_artwork), and the paired LML#478 COMPILATION_ARTIST_* constants — including the label rung as-is, since the #687 gate was dropped and its fix lands against artwork.py afterward. ROWLESS_NO_ALBUM_CONFIDENCE moves to lookup/rowless.py (it is the row-less confidence constant; both new modules consume it, and rowless imports neither, so no validation<->artwork cross-import and no cycle). Orchestrator keeps consumer imports so perform_lookup-driven patch targets (lookup.orchestrator.fetch_artwork_for_items, .filter_results_by_track_validation) keep intercepting; direct-call tests and scripts/measure_artwork_match_floor.py retarget to the new modules. Per the PR-4b green-bisect rule, all import/patch rewires ride this commit so it is independently green.
